requirements.txt
Flask==2.0.3
flask_session==0.4.0
flask_wtf==0.15.1
pymongo==4.7.2
python-dotenv==1.0.1
Werkzeug==2.0.3
pandas
